fix: handle edge cases with insertCompositionText
Sometimes we get multiple characterData mutations for a single composition event, or we get multiple mutations that are not just characterData mutations. In this scenario, we need to manually handle the insertCompositionText event rather than trying to reconcile the DOM from the mutations because it is nearly impossible to make sense of the mutations.